+++ src/ResourceManager/Network/Commands.Network/ApplicationGateway/SslCertificate/SetAzureApplicationGatewaySslCertificateCommand.cs
@@ -39,10 +39,8 @@ namespace Microsoft.Azure.Commands.Network
             {
                 throw new ArgumentException("Ssl certificate with the specified name does not exist");
             }
-
-#pragma warning disable 0618
-            X509Certificate2 cert = new X509Certificate2(CertificateFile, Password, X509KeyStorageFlags.Exportable);
-#pragma warning restore 0618
+            
+            X509Certificate2 cert = new X509Certificate2(this.CertificateFile, this.Password, X509KeyStorageFlags.Exportable);
 
             var newSslCertificate = base.NewObject();