+++ src/ResourceManager/RecoveryServices.Backup/Commands.RecoveryServices.Backup.Test/ScenarioTests/TestsBase.cs
@@ -32,6 +32,11 @@ using RecoveryServicesNS = Microsoft.Azure.Management.RecoveryServices;
 using ResourceManagementNS = Microsoft.Azure.Management.Resources;
 using ResourceManagementRestNS = Microsoft.Azure.Management.ResourceManager;
 using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
+using StorageMgmtNS = Microsoft.Azure.Management.Storage;
+using NetworkMgmtNS = Microsoft.Azure.Management.Network;
+using ComputeMgmtNS = Microsoft.Azure.Management.Compute;
+using System.Collections.ObjectModel;
+using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
 {
@@ -49,6 +54,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
         public ResourceManagementRestNS.ResourceManagementClient RmRestClient { get; private set; }
 
         public HyakRmNS.ResourceManagementClient HyakRmClient { get; private set; }
+
+        public StorageMgmtNS.StorageManagementClient StorageClient { get; private set; }
+
+        public NetworkMgmtNS.NetworkManagementClient NetworkManagementClient { get; private set; }
+
+        public ComputeMgmtNS.ComputeManagementClient ComputeManagementClient { get; private set; }
 
         protected string ResourceNamespace { get; private set; }
 
@@ -70,7 +81,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
         {
             ResourceNamespace = resourceNamespace;
         }
-
+        
         protected void SetupManagementClients(MockContext context)
         {
             RsBackupClient = GetRsBackupClient(context);
@@ -79,12 +90,37 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
             RmRestClient = GetRmRestClient(context);
             HyakRmClient = GetHyakRmClient(context);
 
+            StorageClient = GetStorageManagementClient(context);
+            NetworkManagementClient = GetNetworkManagementClient(context);
+            ComputeManagementClient = GetComputeManagementClient(context);
+
             helper.SetupManagementClients(
                 RsBackupClient,
                 RsClient,
                 RmClient,
                 RmRestClient,
-                HyakRmClient);
+                HyakRmClient,
+                StorageClient,
+                NetworkManagementClient,
+                ComputeManagementClient);
+        }
+
+        private StorageMgmtNS.StorageManagementClient GetStorageManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<StorageMgmtNS.StorageManagementClient>(
+                TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private NetworkMgmtNS.NetworkManagementClient GetNetworkManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<NetworkMgmtNS.NetworkManagementClient>(
+                TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private ComputeMgmtNS.ComputeManagementClient GetComputeManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<ComputeMgmtNS.ComputeManagementClient>(
+                TestEnvironmentFactory.GetTestEnvironment());
         }
 
         private ResourceManagementNS.ResourceManagementClient GetRmClient()
@@ -105,12 +141,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
                 TestEnvironmentFactory.GetTestEnvironment());
         }
 
-        public void RunPsTest(PsBackupProviderTypes providerType, params string[] scripts)
+        public Collection<PSObject> RunPsTest(PsBackupProviderTypes providerType, params string[] scripts)
         {
             var callingClassType = TestUtilities.GetCallingClass(2);
             var mockName = TestUtilities.GetCurrentMethodName(2);
 
-            RunPsTestWorkflow(
+            return RunPsTestWorkflow(
                 providerType,
                 () => scripts,
                 // no custom initializer
@@ -121,7 +157,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
                 mockName);
         }
 
-        public void RunPsTestWorkflow(
+        public Collection<PSObject> RunPsTestWorkflow(
             PsBackupProviderTypes providerType,
             Func<string[]> scriptBuilder,
             Action<LegacyTest.CSMTestEnvironmentFactory> initialize,
@@ -177,6 +213,13 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
                 modules.Add(rmProfileModule);
                 modules.Add(rmModulePath);
                 modules.Add(recoveryServicesModulePath);
+                modules.Add(helper.RMResourceModule);
+                modules.Add(helper.RMStorageDataPlaneModule);
+                modules.Add(helper.RMStorageModule);
+                modules.Add(helper.GetRMModulePath("AzureRM.Compute.psd1"));
+                modules.Add(helper.GetRMModulePath("AzureRM.Network.psd1"));
+                modules.Add("AzureRM.Storage.ps1");
+                modules.Add("AzureRM.Resources.ps1");
 
                 helper.SetupModules(AzureModule.AzureResourceManager, modules.ToArray());
 
@@ -188,7 +231,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
 
                         if (psScripts != null)
                         {
-                            helper.RunPowerShellTest(psScripts);
+                            return helper.RunPowerShellTest(psScripts);
                         }
                     }
                 }
@@ -200,6 +243,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Test.ScenarioTests
                     }
                 }
             }
+
+            return null;
         }
 
         private RecoveryServicesNS.RecoveryServicesClient GetRsClient(MockContext context)